fix: require minimal encoding
ULEB128 technically allows any number trailing "MSB" zeros. However,
this means that there are multiple ways to encode a single number. This
is usually fine but can lead to security issues when the user assumes
that a varint will round-trip and/or that there's exactly one encoding
for some varint.

For example, code may check "is this a CID of type X" by matching the
prefix against a known-byte string. This check only works when there's
exactly one way to encode a varint.

As of https://github.com/multiformats/unsigned-varint/pull/19, the
multiformats unsigned varint spec requires minimal encoding.